Add a method page for editing method
Add a method page in recipe for editing method. Ingredients and method are now edited in separate pages.
