+++ Frontend/App.js
@@ -3,6 +3,7 @@ import Amplify,{ Auth } from '@aws-amplify/core';
 import Fridge from './components/Fridge/Fridge.js';
 import Recipe from './components/Recipe/Recipe.js';
 import RecipeModal from './components/Recipe/RecipeModal.js';
+import MethodModal from './components/Recipe/MethodModal.js';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { NavigationContainer } from '@react-navigation/native';
 export default function App() {
@@ -33,6 +34,7 @@ export default function App() {
         <Stack.Screen name="Fridge" component={Fridge} />
         <Stack.Screen name="Recipe" component={Recipe} />
         <Stack.Screen name="Modify Recipe" component={RecipeModal}/>
+        <Stack.Screen name="Method" component={MethodModal}/>
       </Stack.Navigator>
     </NavigationContainer>
   );
